Danger commit. FastDoubleConverter. Need to Test. Woraround - just replace body by double.Parse.

+++ PoloniexClient/Bittrex/BittrexAccountBalanceInfo.cs
@@ -8,8 +8,8 @@ using System.Threading.Tasks;
 namespace CryptoMarketClient.Bittrex {
     public class BittrexAccountBalanceInfo : BalanceBase {
         public override string Exchange => "Bittrex";
-        public decimal Balance { get; set; }
-        public decimal Pending { get; set; }
+        public double Balance { get; set; }
+        public double Pending { get; set; }
         public bool Requested { get; set; }
         public string Uuid { get; set; }
     }
