+++ PoloniexClient/Bittrex/BittrexOrderInfo.cs
@@ -7,12 +7,12 @@ namespace CryptoMarketClient.Bittrex {
 		public string OrderUuid { get; set; }
 		public string Exchange { get; set; }
         public OrderType OrderType { get; set; }
-        public decimal Quantity { get; set; }
-        public decimal QuantityRemaining { get; set; }
-        public decimal Limit { get; set; }
-        public decimal CommissionPaid { get; set; }
-        public decimal Price { get; set; }
-        public decimal PricePerUnit { get; set; }
+        public double Quantity { get; set; }
+        public double QuantityRemaining { get; set; }
+        public double Limit { get; set; }
+        public double CommissionPaid { get; set; }
+        public double Price { get; set; }
+        public double PricePerUnit { get; set; }
         public DateTime Opened { get; set; }
         public DateTime Closed { get; set; }
         public bool CancelInitiated { get; set; }
