+++ PoloniexClient/Common/BalanceBase.cs
@@ -8,15 +8,15 @@ namespace CryptoMarketClient.Common {
     public abstract class BalanceBase {
         public abstract string Exchange { get; }
         public string Currency { get; set; }
-        public decimal Available { get; set; }
-        public decimal LastAvailable { get; set; }
-        public decimal OnOrders { get; set; }
-        public decimal BtcValue { get; set; }
+        public double Available { get; set; }
+        public double LastAvailable { get; set; }
+        public double OnOrders { get; set; }
+        public double BtcValue { get; set; }
         public string DepositAddress { get; set; }
-        public decimal DepositChanged {
+        public double DepositChanged {
             get {
-                decimal max = Math.Max(Available, LastAvailable);
-                decimal delta = Math.Abs(Available - LastAvailable);
+                double max = Math.Max(Available, LastAvailable);
+                double delta = Math.Abs(Available - LastAvailable);
                 if(max == 0)
                     return 0;
                 return (delta / max);
